Add corrector

+++ src/main/scala/app/models/pixel/GrayscalePixel.scala
@@ -1,6 +1,6 @@
 package app.models.pixel
 
-case class GrayscalePixel(value: Double) extends Pixel {
+case class GrayscalePixel(value: Int) extends Pixel {
   if (value < 0 || value > 255)
     throw new IllegalArgumentException("Value must be between 0 and 255.")
 }
